Fixed the warnings given after the compiling frame.cpp
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4053 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/frame.cpp
@@ -2824,7 +2824,7 @@ void Frame::onListCtrlSelect( wxListEvent &e )
                 wxLIST_STATE_SELECTED );
             while( item != -1 )
             {
-                if ( listCtrlAttr->GetItemData( item ) != static_cast < size_t > (e.GetIndex()) )
+                if ( listCtrlAttr->GetItemData( item ) != e.GetIndex() )
                 {
                     listCtrlAttr->SetItemState(
                         item, 
@@ -2992,7 +2992,7 @@ void Frame::onListCtrlSelect( wxListEvent &e )
                 wxLIST_STATE_SELECTED );
             while( item != -1 )
             {
-                if ( listCtrlClust->GetItemData( item ) != static_cast < size_t > (e.GetIndex()) )
+                if ( listCtrlClust->GetItemData( item ) != e.GetIndex() )
                 {
                     listCtrlClust->SetItemState(
                         item, 
@@ -3021,7 +3021,7 @@ void Frame::onListCtrlSelect( wxListEvent &e )
                 wxLIST_STATE_SELECTED );
             while( item != -1 )
             {
-                if ( listCtrlClust->GetItemData( item ) != static_cast < size_t > (e.GetIndex()) )
+                if ( listCtrlClust->GetItemData( item ) != e.GetIndex() )
                 {
                     if ( selCount > 2 )
                     {
